IGNITE-14183 Test moved to other package.
